docs(Form): move stories of form components to own folder
