+++ packages/admin-stories/src/docs/form/components/stories/FormSection.stories.tsx
@@ -2,7 +2,7 @@ import { Field, FinalForm, FinalFormInput, FormSection } from "@comet/admin";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../../apollo-story.decorator";
+import { apolloStoryDecorator } from "../../../../apollo-story.decorator";
 
 storiesOf("stories/Form/Components", module)
     .addDecorator(apolloStoryDecorator())
